+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - fixed bor, result has same number of bits as its first argument
 - fixed bxor, result has same number of bits as its first argument
 - fixed contrast, 1.2.3.0/24 is really to the left of 1.2.4.0/24
+- IPv6 addresses in lowercase
 
 # 0.1.0
 - initial version